Proper cleanup on forced shutdown. (#4)
The server process didn't exit cleanly when it was killed by the OS when shutting down or rebooting.

This caused the DNS servers to still be set to HostsBeGone ( 127.0.0.1 ) in the system configuration which broke internet until they were cleared out or HostsBeGone is started.

More proper handling of kill signals has been implemented. It istens for `SIGTERM` and `SIGHUP` ( used on Windows systems ) so it can take proper cleanup actions when the system started shutdown/reboot sequence.